+++ docs/CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+* Rename FPSCounter, FPSCounterBundle, FPSCounterSystem to FpsCounter, FpsCounterBundle, FpsCounterSystem. ([#1719])
+
+[#1719]: https://github.com/amethyst/amethyst/pull/1719
+
 ## [0.11.0] - 2019-06
 
 ### Added